guess section cleanup

+++ interpolate.py
@@ -529,7 +529,7 @@ class SmoothSpline(PPform):
         #ndy = y.ndim
         szy = y.shape
 
-        nd = prod(szy[:-1])
+        nd = prod(szy[:-1]).astype(np.int)
         ny = szy[-1]
 
         if n < 2:
